Fix the CI failures on main (#227)
* Fix the CI failures on main

gcc -O3 on pem.c, MSVC uninitialised const statics in sha3 and unique,
mingw's interface macro and signbit/isnan macros, cosmopolitan's close
and write clashing with test helpers, gcc 15's unterminated string
check in pem_test, clang's enum-float check in the smhasher test, a
tidy branch-clone in netip, and a leak in the maps docs example.

* net/netip: zero the address before unmarshal to quiet gcc

The cmake Release build on ubuntu-latest flags addr as maybe
uninitialized in the AddrPort and Prefix unmarshal paths.

* ci: give s390x time for its tests, and keep Dim(-0, 0) at +0 on MSVC

+++ tests/maphash_smhasher_test.c
@@ -340,13 +340,13 @@ static void avalanche_test1(TestingT *t, int nbytes) {
 
     /* The bound Go works out: how many standard deviations make every one of
      * the n*64 cells land inside with probability .9999, times 11 for slack. */
-    double big_n = (double)n * HASH_SIZE;
+    double big_n = (double)n * (double)HASH_SIZE;
     int tenths = 0;
     while (math_pow(math_erf(tenths * .1 / math_sqrt(2)), big_n) < .9999)
         tenths++;
     double c = tenths * .1 * 11.0;
-    double mean = .5 * REP;
-    double stddev = .5 * math_sqrt(REP);
+    double mean = .5 * (double)REP;
+    double stddev = .5 * math_sqrt((double)REP);
     Int low = (Int)(mean - c * stddev);
     Int high = (Int)(mean + c * stddev);
     for (int i = 0; i < n; i++) {
